version=12.2.RC2_dev
# bug fix: command line -M sparshui  not working



Former-commit-id: aa81c4493f135ec6c8bbc4d6f55ff84a46f69a29

+++ Jmol/src/org/jmol/adapter/readers/cifpdb/PdbReader.java
@@ -826,8 +826,8 @@ Polyproline 10
    */
   
   private void structure() {
-    int structureType = 0;
-    int substructureType = 0;
+    byte structureType = 0;
+    byte substructureType = 0;
     int startChainIDIndex;
     int startIndex;
     int endChainIDIndex;
